Move all the Table related components to Table component instead
